+++ src/main/java/tech/showierdata/pickaxe/mixin/ChatHudMixin.java
@@ -1,6 +1,7 @@
 package tech.showierdata.pickaxe.mixin;
 
 import java.util.List;
+import java.util.ListIterator;
 
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
@@ -13,21 +14,18 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 import net.minecraft.text.Text;
 import tech.showierdata.pickaxe.config.Options;
-import tech.showierdata.pickaxe.hook.ChatHudHook;
-import tech.showierdata.pickaxe.hook.IChatHudHook;
+import tech.showierdata.pickaxe.server.Regexs;
 import net.minecraft.client.gui.hud.ChatHud;
 import net.minecraft.client.gui.hud.ChatHudLine;
 import net.minecraft.client.gui.hud.MessageIndicator;
 import net.minecraft.network.message.MessageSignatureData;
 
 @Mixin(ChatHud.class)
-public abstract class ChatHudMixin implements IChatHudHook {
-    /*
-     * Uses an interface to take references to the ChatHud
-     * That interface connects to the hook which allows changes to be made to the Chat Hud outside of Mixin
-     */
+public abstract class ChatHudMixin {
     @Unique
-    private final ChatHudHook chatHudHook = new ChatHudHook(this);
+    private Text prevText = null;
+    @Unique
+    private int count = 1;
 
     @Shadow @Final private List<ChatHudLine> messages;
 
@@ -51,27 +49,49 @@ public abstract class ChatHudMixin implements IChatHudHook {
          */
         if (refreshing) return message;
 
-        return chatHudHook.compactChatMessage(message);
+        return compactChatMessage(message);
+    }
+
+    public void removeMessage(Text originalMessage) {
+        ListIterator<ChatHudLine> iterator = messages.listIterator();
+        while (iterator.hasNext()) {
+            ChatHudLine chatHudLine = iterator.next();
+
+            // Remove previously stacked versions too
+            Text contentWithoutOccurrences = Regexs.removeStackMods(chatHudLine.content());
+            Text textWithoutOccurrences = Regexs.removeStackMods(originalMessage);
+
+            if (contentWithoutOccurrences.equals(textWithoutOccurrences)) {
+                iterator.remove();
+                reset();
+
+                return;
+            }
+        }
+    }
+
+    public Text compactChatMessage(Text message) {
+        // Timestamps are removed to compare texts (otherwise none would match)
+        Text withoutTimestamps = Regexs.removeTimestamps(message);
+
+        Text prevMessage = prevText;
+        prevText = withoutTimestamps;
+
+        // Return if this is new message.
+        if (!withoutTimestamps.equals(prevMessage)) {
+            count = 1;
+            return message;
+        }
+
+        removeMessage(message);
+
+        this.count++;
+        return message.copy().append(String.format(" §8[§bx%s§8]", count));
     }
 
     @Inject(method = "clear", at = @At("RETURN"))
     private void onClear(boolean clearHistory, CallbackInfo info) {
-        this.chatHudHook.clear();
+        prevText = null;
+        count = 1;
     }
-
-    @Override
-    public List<ChatHudLine> getMessages() {
-        return this.messages;
-    }
-
-    @Override
-    public void refreshMessages() {
-        this.reset();
-    }
-
-    @Override
-    public void clear() {
-        this.clear(false);
-    }
-
 }
